opennotes-cli/pyproject.toml
[project]
name = "opennotes-cli"
version = "0.1.0"
description = "CLI for interacting with the OpenNotes server API"
requires-python = ">=3.11"
dependencies = [
    "attrs>=23.2.0",
    "click>=8.1",
    "httpx>=0.27",
    "pydantic>=2.5.0",
    "python-dateutil>=2.8.0",
    "rich>=13.0",
    "trogon>=0.5",
]

[project.scripts]
opennotes = "opennotes_cli.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/opennotes_cli"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[dependency-groups]
dev = [
    "pytest>=7.0",
]
